showSleep.vueとlog.vueをマージ
